functions/requirements.txt
firebase-functions
firebase-admin
google-cloud-firestore
google-generativeai>=0.8.3
